requirements.txt
apache-airflow==2.5.0
pandas==1.4.2
requests==2.26.0
psycopg2-binary==2.9.3
sqlalchemy==1.4.36